Fix issue #1
Problem:

For some environments in symmetric molecules
power spectra of a given L are fully 0
(e.g. odd Ls for an atom in the origin of C_i).
Thus normalization
`vals /= np.sqrt(np.linalg.norm(vals @ vals.T))`
leads to NaNs instead of 0s.
If a molecule is almost symmetric, the norm is ~0,
and the "normalized" power spectra is basically random.

Solution:

Add the `min_norm` argument that allows to zero out
all the power spectra with `norm < min_norm`.
Depending on the application, it can be set to a small number
or to 0 exactly (to treat almost symmetric molecule as asymmetric)